requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pymupdf>=1.24
pydantic>=2.7
python-multipart>=0.0.9
PyYAML>=6.0
regex>=2024.5.15
reportlab>=4.2
python-docx>=1.1.2
